Move the select area and structure into the base class.
